scripts/requirements.txt
scikit-learn
nltk
pinecone
python-dotenv